requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
